requirements.txt
pulumi>=3.0.0,<4.0.0
google-auth
requests